Remove output buffer size computation
It is not clear that the additional allocation is more expensive than
doing the serialization twice.

Moreover, the computation was incorrect for json (and text), where
the wrong type of archive was used, resulting for instance in
considering all ints to use 4 bytes, which is too much for small numbers
and not enough for big ones. If the computed size was too big, then the
array sent would contain trailing zeroes, possibly messing things up on
the receiving end. If the computed size is not big enough, a realloc
would be performed, but the new buffer not used, resulting in a
truncated result.