Fix excessive install
The install command in FindRPCGenerator would install the whole
${output_path} directory, which is usually common to all idl projects.

This resulted in each file in there to be installed as many times as
there are idl projects, a useless slowdown, and worse that would make
the files from test projects arrive in the package.

Also separate the check_sums by project.